fix: Clear search chip highlight when opening CMD+F
